Extract methods for Cart management into Spree::CartMethods module
