+++ frontend/app/controllers/concerns/spree/cart_methods.rb
@@ -1,0 +1,24 @@
+module Spree
+  module CartMethods
+    extend ActiveSupport::Concern
+
+    included do
+      before_action :set_current_order
+      before_action :check_authorization
+    end
+
+    private
+
+    def assign_order_with_lock
+      @order = current_order(lock: true)
+      unless @order
+        flash[:error] = Spree.t(:order_not_found)
+        redirect_to spree.root_path and return
+      end
+    end
+
+    def check_authorization
+      authorize! :edit, current_order, cookies.signed[:token]
+    end
+  end
+end
